requirements.txt
flask==1.1.2
requests==2.25.1
python-dotenv==0.15.0
transformers==4.5.1
torch==1.8.1
numpy==1.20.2
pandas==1.2.4
scikit-learn==0.24.1
streamlit
pydantic